ehr-backend/requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
gunicorn>=21.2,<21.3
python-dotenv>=1.0,<2.0
